+++ payment/src/main/java/org/killbill/billing/payment/config/MultiTenantPaymentConfig.java
@@ -37,7 +37,7 @@ public class MultiTenantPaymentConfig extends MultiTenantConfigBase implements P
 
     @Inject
     public MultiTenantPaymentConfig(@Named(KillBillModule.STATIC_CONFIG) final PaymentConfig staticConfig, final CacheConfig cacheConfig) {
-        super(cacheConfig);
+        super(staticConfig, cacheConfig);
         this.staticConfig = staticConfig;
     }
 
@@ -139,20 +139,6 @@ public class MultiTenantPaymentConfig extends MultiTenantConfigBase implements P
         return getPaymentControlPluginNames();
     }
 
-    @Override
-    public List<TimeSpan> getRescheduleIntervalOnLock() {
-        return staticConfig.getRescheduleIntervalOnLock();
-    }
-
-    @Override
-    public List<TimeSpan> getRescheduleIntervalOnLock(final InternalTenantContext tenantContext) {
-
-        final String result = getStringTenantConfig("getRescheduleIntervalOnLock", tenantContext);
-        if (result != null) {
-            return convertToListTimeSpan(result, "getRescheduleIntervalOnLock");
-        }
-        return getRescheduleIntervalOnLock();
-    }
 
     @Override
     public TimeSpan getJanitorRunningRate() {
