+++ util/src/main/java/org/killbill/billing/util/config/definition/KillbillConfig.java
@@ -16,9 +16,28 @@
  */
 package org.killbill.billing.util.config.definition;
 
+import java.util.List;
+
+import org.killbill.billing.callcontext.InternalTenantContext;
+import org.skife.config.Config;
+import org.skife.config.Default;
+import org.skife.config.Description;
+import org.skife.config.Param;
+import org.skife.config.TimeSpan;
+
 /*
  * Marker interface for killbill config files
  */
 public interface KillbillConfig {
+
+    @Config("org.killbill.rescheduleIntervalOnLock")
+    @Default("30s, 1m, 1m, 3m, 3m, 10m")
+    @Description("Tme delay to reschedule an invoice run when lock is held")
+    List<TimeSpan> getRescheduleIntervalOnLock();
+
+    @Config("org.killbill.rescheduleIntervalOnLock")
+    @Default("30s, 1m, 1m, 3m, 3m, 10m")
+    @Description("Tme delay to reschedule an invoice run when lock is held")
+    List<TimeSpan> getRescheduleIntervalOnLock(@Param("dummy") final InternalTenantContext tenantContext);
 
 }
